refactor: use ky
License: MIT
