+++ src/lib/configure.js
@@ -1,8 +1,10 @@
 'use strict'
+/* eslint-env browser */
 
-const { Headers } = require('node-fetch')
+const ky = require('ky-universal')
+const { isBrowser, isWebWorker } = require('ipfs-utils/src/env')
 const { toUri } = require('./multiaddr')
-const pkg = require('../../package.json')
+const errorHandler = require('./error-handler')
 
 // Set default configuration and call create function with them
 module.exports = create => config => {
@@ -16,28 +18,46 @@ module.exports = create => config => {
     config = { ...config }
   }
 
-  config.fetch = config.fetch || require('./fetch').fetch
-
   if (config.protocol || config.host || config.port) {
     const port = config.port ? `:${config.port}` : ''
     config.apiAddr = `${config.protocol || 'http'}://${config.host || 'localhost'}${port}`
   }
 
-  config.apiAddr = (config.apiAddr || 'http://localhost:5001').toString()
-  config.apiAddr = config.apiAddr.startsWith('/')
-    ? toUri(config.apiAddr)
-    : config.apiAddr
+  config.apiAddr = (config.apiAddr || getDefaultApiAddr(config)).toString()
+  config.apiAddr = config.apiAddr.startsWith('/') ? toUri(config.apiAddr) : config.apiAddr
   config.apiPath = config.apiPath || config['api-path'] || '/api/v0'
 
-  if (config.apiPath.endsWith('/')) {
-    config.apiPath = config.apiPath.slice(0, -1)
+  return create({
+    // TODO configure ky to use config.fetch when this is released:
+    // https://github.com/sindresorhus/ky/pull/153
+    ky: ky.extend({
+      prefixUrl: config.apiAddr + config.apiPath,
+      timeout: config.timeout || 60 * 1000,
+      headers: config.headers,
+      hooks: {
+        afterResponse: [errorHandler]
+      }
+    })
+  })
+}
+
+function getDefaultApiAddr ({ protocol, host, port }) {
+  if (isBrowser || isWebWorker) {
+    if (!protocol && !host && !port) { // Use current origin
+      return ''
+    }
+
+    if (!protocol) {
+      protocol = location.protocol.startsWith('http')
+        ? location.protocol.split(':')[0]
+        : 'http'
+    }
+
+    host = host || location.hostname
+    port = port || location.port
+
+    return `${protocol}://${host}${port ? ':' + port : ''}`
   }
 
-  config.headers = new Headers(config.headers)
-
-  if (!config.headers.has('User-Agent')) {
-    config.headers.append('User-Agent', `${pkg.name}/${pkg.version}`)
-  }
-
-  return create(config)
+  return `${protocol || 'http'}://${host || 'localhost'}:${port || 5001}`
 }
