+++ src/pubsub/unsubscribe.js
@@ -3,7 +3,7 @@
 const configure = require('../lib/configure')
 const SubscriptionTracker = require('./subscription-tracker')
 
-module.exports = configure(({ fetch, apiAddr, apiPath, headers }) => {
+module.exports = configure(({ ky }) => {
   const subsTracker = SubscriptionTracker.singleton()
   // eslint-disable-next-line require-await
   return async (topic, handler) => subsTracker.unsubscribe(topic, handler)
